+++ vllm/v1/metrics/loggers.py
@@ -3,23 +3,18 @@
 import logging
 import time
 from abc import ABC, abstractmethod
-from typing import Callable, Optional, Union, cast
+from typing import Callable, Optional
 
 import numpy as np
 import prometheus_client
 
 from vllm.config import SupportsMetricsInfo, VllmConfig
-from vllm.executor.ray_utils import ray
 from vllm.logger import init_logger
 from vllm.v1.core.kv_cache_utils import PrefixCachingMetrics
 from vllm.v1.engine import FinishReason
 from vllm.v1.metrics.stats import IterationStats, SchedulerStats
+from vllm.v1.metrics.wrappers import RayMetrics
 from vllm.v1.spec_decode.metrics import SpecDecodingLogging, SpecDecodingProm
-
-if ray is not None:
-    from ray.util import metrics as ray_metrics
-else:
-    ray_metrics = None
 
 logger = init_logger(__name__)
 
@@ -143,7 +138,7 @@ class LoggingStatLogger(StatLoggerBase):
             self.vllm_config.cache_config.num_gpu_blocks)
 
 
-class Metrics:
+class PrometheusMetrics:
     """
     vLLM uses a multiprocessing-based frontend for the OpenAI server.
     This means that we need to run prometheus_client in multiprocessing mode
@@ -441,119 +436,8 @@ class Metrics:
                 prometheus_client.REGISTRY.unregister(collector)
 
 
-class _RayGaugeWrapper:
-    """Wraps around ray.util.metrics.Gauge to provide same API as
-    prometheus_client.Gauge"""
-
-    def __init__(self,
-                 name: str,
-                 documentation: str = "",
-                 labelnames: Optional[list[str]] = None,
-                 multiprocess_mode: str = ""):
-        del multiprocess_mode
-        labelnames_tuple = tuple(labelnames) if labelnames else None
-        self._gauge = ray_metrics.Gauge(name=name,
-                                        description=documentation,
-                                        tag_keys=labelnames_tuple)
-
-    def labels(self, **labels):
-        self._gauge.set_default_tags(labels)
-        return self
-
-    def set(self, value: Union[int, float]):
-        return self._gauge.set(value)
-
-    def set_to_current_time(self):
-        # ray metrics doesn't have set_to_current time, https://docs.ray.io/en/latest/_modules/ray/util/metrics.html
-        return self._gauge.set(time.time())
-
-
-class _RayCounterWrapper:
-    """Wraps around ray.util.metrics.Counter to provide same API as
-    prometheus_client.Counter"""
-
-    def __init__(self,
-                 name: str,
-                 documentation: str = "",
-                 labelnames: Optional[list[str]] = None):
-        labelnames_tuple = tuple(labelnames) if labelnames else None
-        self._counter = ray_metrics.Counter(name=name,
-                                            description=documentation,
-                                            tag_keys=labelnames_tuple)
-
-    def labels(self, **labels):
-        self._counter.set_default_tags(labels)
-        return self
-
-    def inc(self, value: Union[int, float] = 1.0):
-        if value == 0:
-            return
-        return self._counter.inc(value)
-
-
-class _RayHistogramWrapper:
-    """Wraps around ray.util.metrics.Histogram to provide same API as
-    prometheus_client.Histogram"""
-
-    def __init__(self,
-                 name: str,
-                 documentation: str = "",
-                 labelnames: Optional[list[str]] = None,
-                 buckets: Optional[list[float]] = None):
-        labelnames_tuple = tuple(labelnames) if labelnames else None
-        boundaries = buckets if buckets else []
-        self._histogram = ray_metrics.Histogram(name=name,
-                                                description=documentation,
-                                                tag_keys=labelnames_tuple,
-                                                boundaries=boundaries)
-
-    def labels(self, **labels):
-        self._histogram.set_default_tags(labels)
-        return self
-
-    def observe(self, value: Union[int, float]):
-        return self._histogram.observe(value)
-
-
-class RaySpecDecodingProm(SpecDecodingProm):
-    """
-    RaySpecDecodingProm is used by RayMetrics to log to Ray metrics.
-    Provides the same metrics as SpecDecodingProm but uses Ray's
-    util.metrics library.
-    """
-
-    _counter_cls: type[prometheus_client.Counter] = cast(
-        type[prometheus_client.Counter], _RayCounterWrapper)
-
-
-class RayMetrics(Metrics):
-    """
-    RayMetrics is used by RayPrometheusStatLogger to log to Ray metrics.
-    Provides the same metrics as Metrics but uses Ray's util.metrics library.
-    """
-
-    _gauge_cls: type[prometheus_client.Gauge] = cast(
-        type[prometheus_client.Gauge], _RayGaugeWrapper)
-    _counter_cls: type[prometheus_client.Counter] = cast(
-        type[prometheus_client.Counter], _RayCounterWrapper)
-    _histogram_cls: type[prometheus_client.Histogram] = cast(
-        type[prometheus_client.Histogram], _RayHistogramWrapper)
-    _spec_decoding_cls: type[SpecDecodingProm] = cast(type[SpecDecodingProm],
-                                                      RaySpecDecodingProm)
-
-    def __init__(self, vllm_config: VllmConfig, engine_index: int = 0):
-        if ray_metrics is None:
-            raise ImportError("RayMetrics requires Ray to be installed.")
-        super().__init__(vllm_config, engine_index)
-
-    @staticmethod
-    def _unregister_vllm_metrics():
-        # No-op on purpose
-        pass
-
-
 class PrometheusStatLogger(StatLoggerBase):
-    _metrics_cls = Metrics
+    _metrics_cls = PrometheusMetrics
 
     def __init__(self, vllm_config: VllmConfig, engine_index: int = 0):
         self.vllm_config = vllm_config
